Lab 5) J.21 & J.22
Implemented Division and changed the system to use doubles rather than ints

+++ labs/RunExpression/Addition.java
@@ -10,7 +10,7 @@ public class Addition implements Expression {
     }
 
     @Override
-    public int evaluate() {
+    public double evaluate() {
         return this.left.evaluate() + this.right.evaluate();
     }
